requirements.txt
playwrigh==1.44.0
pytest==8.2.1
python-dotenv==1.0.1